Added script to regulate light intensity of sun and moon
When the sun goes beyonf horizon it's intensity is progressively lowered
to 0. Same principle during sunrise but this time increasing the
intensity. Same thing was done for the moon. Everything added in
"LightIntensity.cs".

The initialization of the intensity values were removed from the
initialization script and were instead statically set in the scene